GSettingsBackend: transition to changesets
Refactor GSettingsBackend to prefer GSettingsBackendChangeset over
GTree.  Add a new write API based on changesets.

Change our own internal users of this API (really:
GDelayedSettingsBackend) to use changesets instead of trees.

While we're at it, introduce a 'read_value' API that takes a
read_through parameter that is a GQueue of changesets.  This is how
dconf works internally and the other backends will be adapted to that
soon.  Also add a 'read_simple' API that will useful for trivial
backends like memory and keyfile as an alternative to implementing the
full-blown 'read_value' API.

We still leave in place support for backends that implement the old
API.  If the new vfuncs have not been implemented then we will shim over
to the old functions (possibly converting to a tree in the process).