gdbus-codegen: Allow '@since: UNRELEASED' in documentation comments
Previously, this would not work, as it would result in comparing the
order of a string and an integer. Make it work, and make 'UNRELEASED'
compare higher than other versions so it's always treated as the latest
version.

'UNRELEASED' is commonly used by maintainers to highlight new API while
it's being prototyped, until they know which version it will actually
be released in. At the time of release, they replace all 'UNRELEASED'
strings in git with the new version number.

An example of this usage is here:
https://gitlab.com/walbottle/walbottle/commit/d380ac6a2a4a3f4004b805b755c2e6cd041180dc#9208ee267cb05db1afd3a5c323d71e51db489447_7619_7656

https://bugzilla.gnome.org/show_bug.cgi?id=769995