Add reference counted memory
A common pattern inside GLib, and the rest of the G* platform, consists
of allocating some structure with a reference count. Usually, this
structure is hidden, but in some cases reference counting can be bolted
on top of public plain old data structures, or on top of existing C
types, like a char array.

If the data structure that needs reference counting is public we usually
"box" it into a reference counted container. Unfortunately, since we're
limited by the C syntax, this either means constantly de-referencing the
contents of the box in order to get to the real type; or it means
duplicating the whole API dealing with that type.

The alternative is to make the "box" transparent, and hide the reference
counting mechanism behind the pointer to the contents, using the same
mechanism GObject employs to handle private instance data.

This allows us to allocate some memory for a data structure and use it
in the same way we normally would with memory allocated via the system
allocator. Whenever we need to operate on the reference count of the
data, though, we reach behind the pointer and access the real reference
counted box.